Move `singer_sdk/internal/about.py -> singer_sdk/about.py`
