+++ singer_sdk/about.py
